requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai>=1.0.0
azure-identity>=1.12.0
azure-search-documents>=11.4.0
httpx>=0.25.0
pymupdf>=1.24.14
openpyxl==3.1.2
chromadb>=0.5.3
rank-bm25==0.2.2
python-multipart==0.0.9
